Queryd

+++ misc/gen_tar_lists.py
@@ -1,3 +1,6 @@
+"""
+ipy misc/gen_tar_lists.py -- --dataset YouCook2
+"""
 import copy
 import json
 import argparse
@@ -5,22 +8,35 @@ from typing import Dict, List, Tuple
 from pathlib import Path
 
 import tqdm
-from typeguard import typechecked
+from beartype import beartype
+from zsvision.zs_utils import load_json_config
+from gen_readme import dataset_paths, model_specs2path
 
-from parse_config import ConfigParser
-from misc.gen_readme import dataset_paths, model_specs2path
 
-
-@typechecked
-def generate_tar_lists(save_dir: Path, experiments: Dict[str, Tuple[str, str]]):
+@beartype
+def generate_tar_lists(
+        save_dir: Path,
+        experiments: Dict[str, Tuple[str, str]],
+        datasets: List[str],
+        refresh: bool,
+):
     all_feat_paths = {}
     for exp_name, (group_id, timestamp) in tqdm.tqdm(experiments.items()):
         rel_path = Path(group_id) / "seed-0" / timestamp / "config.json"
         config_path = Path(save_dir) / "models" / exp_name / rel_path
-        with open(config_path, "r") as f:
-            config = json.load(f)
+        try:
+            with open(config_path, "r") as f:
+                config = json.load(f)
+        except FileNotFoundError:
+            rel_path = Path(group_id) / "seed-1" / timestamp / "config.json"
+            config_path = Path(save_dir) / "models" / exp_name / rel_path
+            with open(config_path, "r") as f:
+                config = json.load(f)
+
         feat_aggregation = config["data_loader"]["args"]["feat_aggregation"]
         dataset_name = exp_name.split("-train")[0]
+        if dataset_name not in [x.lower() for x in datasets]:
+            continue
         if dataset_name not in all_feat_paths:
             all_feat_paths[dataset_name] = set()
         split_names = [config["data_loader"]["args"]["split_name"]]
@@ -37,8 +53,7 @@ def generate_tar_lists(save_dir: Path, experiments: Dict[str, Tuple[str, str]]):
         for key, feat_list in paths["custom_paths"].items():
             for feat_path in feat_list:
                 all_feat_paths[dataset_name].add(root_feat / feat_path)
-        text_paths = [root_feat / rel_path for rel_path in
-                      paths["text_feat_paths"][text_feat].values()]
+        text_paths = [root_feat / paths["text_feat_paths"][text_feat]]
         all_feat_paths[dataset_name].update(set(text_paths))
         all_feat_paths[dataset_name].add(root_feat / paths["raw_captions_path"])
         if "dict_youtube_mapping_path" in paths:
@@ -52,13 +67,17 @@ def generate_tar_lists(save_dir: Path, experiments: Dict[str, Tuple[str, str]]):
     for dataset_name, paths in all_feat_paths.items():
         tar_include_list = Path("misc") / "datasets" / dataset_name / "tar_include.txt"
         tar_include_list.parent.mkdir(exist_ok=True, parents=True)
+        if tar_include_list.exists() and not refresh:
+            print(f"Found existing tar include list at {tar_include_list}, skipping...")
+            continue
         with open(tar_include_list, "w") as f:
             for path in sorted(paths):
-                print(f"Writing {path} to {tar_include_list}")
-                f.write(f"{path}\n")
+                if "aggregated_speech" not in str(path):
+                    print(f"Writing {path} to {tar_include_list}")
+                    f.write(f"{path}\n")
 
 
-@typechecked
+@beartype
 def generate_tar_lists_for_challenge(
         refresh: bool,
         datasets: List[str],
@@ -84,7 +103,7 @@ def generate_tar_lists_for_challenge(
         print(f"Found {len(rel_paths)} files in {src_folder}")
         fname = f"data_loader_{dataset.lower()}.json"
         config_path = Path("configs") / "cvpr2020-challenge" / fname
-        config = ConfigParser.load_config(config_path)
+        config = load_json_config(config_path)
 
         keep = set(config["experts"]["modalities"])
         feat_aggregation = config["data_loader"]["args"]["feat_aggregation"]
@@ -135,16 +154,15 @@ def generate_tar_lists_for_challenge(
                 f.write(f"{path}\n")
 
         # select video paths
-        if challenge_phase == "public_server_val":
-            video_dir = src_folder / "videos"
-            video_paths = [x for x in rel_paths if str(x).startswith(str(video_dir))
-                        and x.is_file()]
-            print(f"[{dataset}] Found {len(video_paths)} video paths")
+        video_dir = src_folder / "videos"
+        video_paths = [x for x in rel_paths if str(x).startswith(str(video_dir))
+                       and x.is_file()]
+        print(f"[{dataset}] Found {len(video_paths)} video paths")
 
-            with open(video_tar_include_list, "w") as f:
-                print(f"Writing video paths to {video_tar_include_list}")
-                for path in sorted(video_paths):
-                    f.write(f"{path}\n")
+        with open(video_tar_include_list, "w") as f:
+            print(f"Writing video paths to {video_tar_include_list}")
+            for path in sorted(video_paths):
+                f.write(f"{path}\n")
 
 
 def main():
@@ -158,7 +176,8 @@ def main():
     parser.add_argument("--challenge_phase", default="public_server_val",
                         choices=["public_server_val", "public_server_test"])
     parser.add_argument("--datasets", nargs="+",
-                        default=["MSRVTT", "MSVD", "DiDeMo", "activity-net", "YouCook2"])
+                        default=["MSRVTT", "MSVD", "DiDeMo", "activity-net", "YouCook2",
+                                 "QuerYD", "QuerYDSegments"])
     args = parser.parse_args()
 
     with open(args.experiments_path, "r") as f:
@@ -167,7 +186,9 @@ def main():
     if args.target == "main":
         generate_tar_lists(
             save_dir=args.save_dir,
+            datasets=args.datasets,
             experiments=experiments,
+            refresh=args.refresh,
         )
     elif args.target == "cvpr2020_challenge":
         generate_tar_lists_for_challenge(
